requirements.txt
# LLM & Embeddings
groq
google-generativeai
sentence-transformers
torch

# Database
pymongo[srv]

# PDF Processing
PyMuPDF
pdfplumber

# Backend
fastapi
uvicorn
python-multipart

# Utilities
python-dotenv
pydantic
httpx
requests
langchain-text-splitters
firecrawl-py
